% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingTreatment.R
\name{mice_treatment}
\alias{mice_treatment}
\title{Multivariate Imputation by Chained Equations}
\usage{
mice_treatment(data, dv, method)
}
\arguments{
\item{data}{Any data frame tat needs to be treated for Missing data}

\item{dv}{The dependant variable that is to be ignored from the given data while treating missing values}

\item{method}{Can be either a single string, or a vector of strings with length length(blocks), specifying the imputation method to be used for each column in data. If specified as a single string, the same method will be used for all blocks.}
}
\value{
Returns a dataframe treated for missing values
}
\description{
The mice package implements a method to deal with missing data. The package creates multiple imputations (replacement values) for multivariate missing data. The method is based on Fully Conditional Specification, where each incomplete variable is imputed by a separate model. The MICE algorithm can impute mixes of continuous, binary, unordered categorical and ordered categorical data. In addition, MICE can impute continuous two-level data, and maintain consistency between imputations by means of passive imputation. Many diagnostic plots are implemented to inspect the quality of the imputations. Generates Multivariate Imputations by Chained Equations (MICE)
}
