% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingTreatment.R
\name{missing_treatments}
\alias{missing_treatments}
\title{Treats missing values in a given dataset using the best possible technique}
\usage{
missing_treatments(data, dv)
}
\arguments{
\item{data}{Any data frame that has to treated for missing data}

\item{dv}{Dependent variable in the given dataset in order to eliminate that field while treating missing data}
}
\value{
Returns a list of 3 objects:
\describe{
\item{missing_treated}{Missing value treated Dataset}
\item{fit_model}{A list of model fit files for each of the techniques used}
\item{model_perf_metrics}{Performance metrics for eac of the intermediate datasets created}
}
}
\description{
missing_treatments initially eliminates all the columns having more than 70% missing data. Different missing value treatment techiniques considered  are as follows:
\itemize{
\item Zero Imputation
\item Min Imputation
\item Max Imputation
\item Mean Imputation
\item Median Imputation
\item Mode Imputation
\item Multiple Imputation - Random Forest
\item Multiple Imputation - Predictive Mean Matching
}
Of all these techniques, Best technique is chosed based on the performance of different datasets on a linear/logistic model.
}
\details{
Takes in a data frame and performs the best possible missing value treatment to each of the columns in the data frame
}
